requirements.txt
tensorflow-cpu
streamlit
numpy
gdown